Changed timing to use QueryPerformanceCounter instead of GetTickCount. Much more accurate now.

+++ engine/include/ElapsedTimer.h
@@ -61,7 +61,7 @@ public:
 	float getTime() const;
 
 private:
-	unsigned long m_startTime;
+	float m_startTime;
 };
 
 }
